Add: [CDN] code and config to generate required files for CDN
This is including the workflow scripts to automate the CDN
generator.

The CDN generator includes:
 - index.html: an index file (as S3 / CloudFront / CloudFlare
               cannot do this for us)
 - folders.yaml: to be able to script walking the CDN
 - latest.yaml: a short summary for the latest release of a
                category
 - manifest.yaml: manifest file per release

This can be generated for everything, or only for a single (new)
release. The first is a rebuild, the second is for when a new
release is pushed to the CDN. In the second case the bare minimum
will be done to come to the same result as a full rebuild would
have done.

This requires read access to the S3 via the AWS API.